Reconstruct the code structure
